+++ src/main/java/chess/domain/Point.java
@@ -1,4 +1,4 @@
-package chess;
+package chess.domain;
 
 public class Point {
