+++ src/test/java/chess/domain/PointTest.java
@@ -1,4 +1,4 @@
-package chess;
+package chess.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
